\encoding{UTF-8}
\name{birewire.load.dsg}

\alias{birewire.load.dsg}
\title{Read a SIF file from a given path}
\description{The routine reads a SIF file and return a R table.} 
\usage{birewire.load.dsg(path)}
\arguments{
 \item{path}{Path to the SIF  file.}
 } 


\value{
A R table that can be transfomred into a dsg using \code{birewire.induced.bipartite}}
\references{
 Iorio, F. and Gobbi, A.  and  Saez-Rodriguez, J. (2015) \emph{XXX} YYY.\cr
  }

\keyword{directed graph, rewire, pathway, signaling}
