\encoding{UTF-8}
\name{get.data.frame.from.incidence}

\alias{get.data.frame.from.incidence}
\title{Internal function for the generation of a SIF file of a given dsg.}
\description{Internal routine for \code{birewire.build.dsg}} 
\usage{get.data.frame.from.incidence(table,sign)}
\arguments{
 \item{table}{An incidence matrix to be converted in SIF format }
  \item{sign}{The sign of the relation that the incidence matrix encodes }

 } 


\value{
A SIF-like dataframe used to generate the complete SIF file in \code{birewire.build.dsg}.}
\references{
 Iorio, F. and Gobbi, A.  and  Saez-Rodriguez, J. (2015) \emph{XXX} YYY.\cr
  }

\keyword{directed graph, rewire, pathway, signaling}
