\encoding{UTF-8}
\name{simplify.table}

\alias{simplify.table}
\title{Internal function for the generation of ondiced bipartite graph starting froma a given dsg.}
\description{Internal routine for \code{birewire.induced.bipartite}} 
\usage{simplify.table(df)}
\arguments{
 \item{df}{An incidence matrix describing a bipartite graph that must be simplify in order to have only nodes with positive degree. }

 } 


\value{
An incidence matrix simplified that can be used in \code{birewire.rewire.bipartite}.}
\references{
 Iorio, F. and Gobbi, A.  \emph{In preparation} In preparation.\cr
  }

\keyword{directed graph, rewire, pathway, signaling}
