\encoding{UTF-8}
\name{birewire.load.dsg}

\alias{birewire.load.dsg}
\title{Read a SIF file from a given path}
\description{The routine reads a SIF file and return a R table.} 
\usage{birewire.load.dsg(path)}
\arguments{
 \item{path}{Path to the SIF  file.}
 } 


\value{
A R table that can be transfomred into a dsg using \code{\link{birewire.induced.bipartite}}}


\keyword{directed graph, rewire, pathway, signaling}
